不要な変更を削除
