--- src/app/admin/credentials/components/DateSelector.tsx
@@ -1,7 +1,0 @@
-export default function DateSelector() {
-    return (
-        <div>
-            <h2>日付を選択</h2>
-        </div>
-    );
-}
