+++ src/app/admin/credentials/issue/page.tsx
@@ -1,10 +1,10 @@
-import TicketSelector from "../components/TicketSelector";
+import OpportunitySelector from "../components/OpportunitySelector";
 
 export default function SelectOpportunity() {
     return (
         <div className="p-4 space-y-8 max-w-2xl mx-auto">
             <div className="space-y-4">
-                <TicketSelector />
+                <OpportunitySelector />
             </div>
         </div>
     );
